+++ examples/memory-todo-agent/CHANGELOG.md
@@ -1,5 +1,60 @@
 # memory-todo-agent
 
+## 0.1.0-alpha.21
+
+### Patch Changes
+
+- Updated dependencies [d5fccfb]
+  - @mastra/core@0.2.0-alpha.109
+  - @mastra/memory@0.1.0-alpha.91
+
+## 0.1.0-alpha.20
+
+### Patch Changes
+
+- Updated dependencies [5ee67d3]
+- Updated dependencies [95a4697]
+  - @mastra/core@0.2.0-alpha.108
+  - @mastra/memory@0.1.0-alpha.90
+
+## 0.1.0-alpha.19
+
+### Patch Changes
+
+- Updated dependencies [66a5392]
+  - @mastra/core@0.2.0-alpha.107
+  - @mastra/memory@0.1.0-alpha.89
+
+## 0.1.0-alpha.18
+
+### Patch Changes
+
+- Updated dependencies [6f2c0f5]
+- Updated dependencies [a8a459a]
+  - @mastra/core@0.2.0-alpha.106
+  - @mastra/memory@0.1.0-alpha.88
+
+## 0.1.0-alpha.17
+
+### Patch Changes
+
+- Updated dependencies [1420ae2]
+- Updated dependencies [99f1847]
+  - @mastra/core@0.2.0-alpha.105
+  - @mastra/memory@0.1.0-alpha.87
+
+## 0.1.0-alpha.16
+
+### Patch Changes
+
+- Updated dependencies [5fdc87c]
+- Updated dependencies [b97ca96]
+- Updated dependencies [72d1990]
+- Updated dependencies [cf6d825]
+- Updated dependencies [10870bc]
+  - @mastra/core@0.2.0-alpha.104
+  - @mastra/memory@0.1.0-alpha.86
+
 ## 0.1.0-alpha.15
 
 ### Patch Changes